document requisition report 95% done

+++ resources/views/back-end/requisition/_project_wise_data_type_wise_document_status_report_details.blade.php
@@ -1,16 +1,18 @@
 <div class="modal-header">
-    <h5 class="modal-title" id="dataTypesDetailsLabel">{!! $result->project_name !!} Required Data Type Details Report</h5>
+    <h5 class="modal-title" id="dataTypesDetailsLabel"><i class="fas fa-file-lines"></i> {!! $result->project_name !!} Required Data Type Details Report</h5>
     <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
 </div>
 <div class="modal-body">
-    @if(isset($result->data_types) && count($result->data_types))
-    <label>Selected Oprations</label>
-    <button class="btn btn-sm btn-outline-success mb-2" onclick="return DataTypeNecessityChange(this,1,{!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fa-solid fa-star"></i> Make Required</button>
-    <button class="btn btn-sm btn-outline-info mb-2" onclick="return DataTypeNecessityChange(this,0,{!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fa-regular fa-star"></i> Make Optional</button>
-    <button class="btn btn-sm btn-outline-danger mb-2" onclick="return DeleteProjectWiseNecessaryDataType({!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fas fa-trash"></i> Delete</button>
+    @if(auth()->user()->hasPermission('project_document_requisition_edit'))
+        @if(isset($result->data_types) && count($result->data_types))
+            <label>Selected Oprations</label>
+            <button class="btn btn-sm btn-outline-success mb-2" onclick="return DataTypeNecessityChange(this,1,{!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fa-solid fa-star"></i> Make Required</button>
+            <button class="btn btn-sm btn-outline-info mb-2" onclick="return DataTypeNecessityChange(this,0,{!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fa-regular fa-star"></i> Make Optional</button>
+            <button class="btn btn-sm btn-outline-danger mb-2" onclick="return DeleteProjectWiseNecessaryDataType({!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fas fa-trash"></i> Delete</button>
+        @endif
+        <button class="btn btn-sm btn-outline-primary mb-2 @if(isset($result->data_types) && count($result->data_types))float-end @endif"  onclick="return ProjectWiseNewDataTypeAdd({!! $result->pdri_id !!},{!! $result->project_id !!},{!! $result->company_id !!},)"><i class="fas fa-plus"></i> Add Data Type</button>
+        <hr>
     @endif
-    <button class="btn btn-sm btn-outline-primary mb-2 @if(isset($result->data_types) && count($result->data_types))float-end @endif"><i class="fas fa-plus"></i> Add Data Type</button>
-    <hr>
     <table @if($result) id="DataTypeTable" @endif class="table table-hover table-striped table-sm" style="width:100%">
         <thead>
         <tr>
@@ -22,9 +24,10 @@
             <th>Data Type</th>
             <th>Necessity</th>
             <th>Deadline</th>
-            <th>Document Status</th>
-            <th>Document Count</th>
-            <th title="Responsible User Count">Responsible Count</th>
+            <th title="Document Status">Status</th>
+            <th title="Document Count">Document's</th>
+            <th title="Responsible Departments">Res. Depts.</th>
+            <th title="Responsible Users">Res. Users</th>
         </tr>
         </thead>
         <tbody>
@@ -39,10 +42,23 @@
                 <td>{!! $row->data_type_name !!}</td>
                 <td >{!! $row->necessity? "<span class='badge bg-success'><i class='fa-solid fa-star'></i> Required</span>": "<span class='badge bg-info'><i class='fa-regular fa-star'></i> Optional</span>"!!}</td>
                 <td>{!! $row->deadline !!}</td>
-                <td>{!! $row->documents?'✅ OK':'❌ Missing' !!}</td>
-                <td><a href="{!! route('uploaded.archive.list.pagination', ['c' => $result->company_id, 't' => $row->data_type_id,'p' => $result->project_id]) !!}" target="_blank">{!! $row->documents??'0' !!}</a></td>
+                <td>{!! $row->documents?'<span class="badge bg-success"><i class="fa-solid fa-circle-check"></i> Ok</span':'<span class="badge bg-danger"> <i class="fa-solid fa-circle-xmark"></i> Missing</span>' !!}</td>
                 <td>
-                    {!! $row->responsible_by_count !!}
+                    @if(auth()->user()->hasPermission("archive_data_list_quick"))
+                        <a href="{!! route('uploaded.archive.list.pagination', ['c' => $result->company_id, 't' => $row->data_type_id,'p' => $result->project_id]) !!}" target="_blank">{!! $row->documents??'0' !!}</a>
+                    @else
+                        {!! $row->documents??'0' !!}
+                    @endif
+                </td>
+                <td>
+                    @if($row->departments)
+                        @foreach($row->departments as $dept)
+                            <span class="badge bg-secondary"> {!! $dept->dept_name !!}</span>
+                        @endforeach
+                    @endif
+                </td>
+                <td>
+                    <a href="#" onclick="return DataTypeWiseResponsibleUserAdd(this,{{$row->req_data_type_id}},{!! $result->company_id !!})">{!! $row->responsible_by_count !!}</a>
 {{--                    @foreach($row->responsible_by as $usr)--}}
 {{--                        <span class="badge bg-secondary"> {!! $usr->name !!} ({!! $usr->employee_id !!})</span>--}}
 {{--                    @endforeach--}}
